restructure to add drawer nav
