--- app/(home)/_layout.tsx
@@ -1,7 +1,0 @@
-import { Stack } from 'expo-router'
-
-export default function AppLayout() {
-  return (
-    <Stack screenOptions={{ headerShown: false }} />
-  )
-}
