+++ app/(tabs)/main/cart/payment.tsx
@@ -1,95 +1,76 @@
 // react native screen with title 
 
 import React, { useEffect, useState } from 'react';
-import { View, Text, StyleSheet, SafeAreaView, TouchableOpacity } from 'react-native';
-import { useAuth } from '../../contexts/AuthContext';
+import { View, Text, StyleSheet, SafeAreaView, ScrollView, TouchableOpacity } from 'react-native';
+import { useAuth } from '../../../contexts/AuthContext';
 import { Button, Image, Input, Switch } from '@rneui/themed';
-import { useProducts } from '../../contexts/ProductContext';
-import BackButton from '../../components/BackButton';
-import { Formik } from 'formik';
-import colors from '../../config/colors';
-import { useCart } from '../../contexts/cart/CartContext';
+import { useProducts } from '../../../contexts/ProductContext';
+import BackButton from '../../../components/BackButton';
+import { Feather } from '@expo/vector-icons';
+import colors from '../../../config/colors';
 import { useRouter } from 'expo-router';
-import { CardField } from '@stripe/stripe-react-native';
+import { useCart } from '../../../contexts/cart/CartContext';
+import { Card } from '../../../types/global';
+import { Formik } from 'formik';
+import PaymentCard from '../../../components/PaymentCard';
 
-export default function Address() {
+export default function Payment() {
+
+  const [saveCard, setSaveCard] = useState(false);
+  const { cards } = useCart()
   const router = useRouter();
-  const [isPrimary, setIsPrimary] = useState(false);
-  const [selected, setSelected] = useState<'card' | 'paypal' | 'applepay'>('card');
-  const { user, session, signOut } = useAuth();
+  const { user } = useAuth();
   const { submitCard } = useCart();
+  const toggleSwitch = () => setSaveCard(previousState => !previousState);
+
+  useEffect(() => {
+    console.log('cards', cards);
+  }, [cards]);
 
   return (
     <SafeAreaView style={styles.container}>
       <View style={{}}>
         <View style={styles.headerContainer}>
-          <Text style={styles.headerTitle}>Add New Card</Text>
+          <Text style={styles.headerTitle}>Payment</Text>
         </View>
         <View style={styles.backButton}>
           <BackButton />
         </View>
       </View>
-      <View style={{
-        marginHorizontal: 20,
-        marginTop: 38,
-        flexDirection: 'row',
-        justifyContent: 'space-between',
-      }}>
-        <View style={[styles.paymentLogo, {
-          backgroundColor: selected === 'card' ? colors.secondary2 : colors.white,
-          borderColor: selected === 'card' ? colors.orange : colors.grey5,
-          borderWidth: 1
-        }]}>
-          <TouchableOpacity
-            onPress={() => {
-              router.push({
-                pathname: '/(tabs)/home/brandProducts',
-              })
-            }}
+      <View>
+        <ScrollView
+          horizontal
+          showsHorizontalScrollIndicator={false}
+          style={{ alignSelf: cards.length > 1 ? undefined : 'center' }}
+        >
+          {/* list of cards, later get from stripe api or db */}
+          <View style={{
+            flexDirection: 'row',
+            justifyContent: 'center',
+          }}
           >
-            <Image source={{ uri: 'https://via.placeholder.com/150' }}
-              style={{
-                width: 26,
-                height: 17,
-                alignItems: 'center',
-              }}
-            />
-          </TouchableOpacity>
-        </View>
-        <View style={styles.paymentLogo}>
-          <TouchableOpacity
-            onPress={() => {
-              router.push({
-                pathname: '/(tabs)/home/brandProducts',
-              })
-            }}
-          >
-            <Image source={{ uri: 'https://via.placeholder.com/150' }}
-              style={{
-                width: 26,
-                height: 17,
-                alignItems: 'center',
-              }}
-            />
-          </TouchableOpacity>
-        </View>
-        <View style={styles.paymentLogo}>
-          <TouchableOpacity
-            onPress={() => {
-              router.push({
-                pathname: '/(tabs)/home/brandProducts',
-              })
-            }}
-          >
-            <Image source={{ uri: 'https://via.placeholder.com/150' }}
-              style={{
-                width: 26,
-                height: 17,
-                alignItems: 'center',
-              }}
-            />
-          </TouchableOpacity>
-        </View>
+            {
+              cards.map((card, index) => (
+                <PaymentCard
+                  key={index}
+                  card={card}
+                  submitCard={submitCard}
+                  router={router}
+                />
+              ))
+            }
+          </View>
+        </ScrollView>
+      </View>
+
+      <View>
+        {/* add new card button */}
+        <TouchableOpacity style={styles.addNewCardButton}
+          onPress={() => router.push('/(tabs)/main/cart/addNewCard')}
+        >
+          <Feather name="plus-square" size={15} color={colors.primary} />
+          <Text style={styles.addNewCardText}>Add New Card</Text>
+        </TouchableOpacity>
       </View>
       <Formik
         initialValues={{
@@ -99,6 +80,7 @@ export default function Address() {
           card_number: '5254 5454 5454 5454',
           exp: '12/22',
           cvv: '7763',
+          save_card: saveCard,
           user_id: user?.id!,
         }}
         onSubmit={async (values) => {
@@ -106,7 +88,7 @@ export default function Address() {
           try {
             // if (error) throw error
             console.log('values submitting CARD', values)
-            submitCard({ ...values }, router, true);
+            submitCard({ ...values }, router);
           } catch (err) {
             console.log('Error', err)
             throw err;
@@ -179,6 +161,15 @@ export default function Address() {
               </View>
             </View>
             {/* submit button */}
+            <View style={styles.toggle}>
+              <Text style={styles.toggleText}>Save card info</Text>
+              <Switch
+                onValueChange={toggleSwitch}
+                value={saveCard}
+                style={styles.toggleSwitch}
+                trackColor={{ false: colors.grey3, true: colors.secondary3 }}
+              />
+            </View>
             <TouchableOpacity
               onPress={handleSubmit}
               style={styles.button}
@@ -212,7 +203,7 @@ const styles = StyleSheet.create({
   backButton: {
     position: 'absolute',
     left: 20,
-    top: 0,
+    top: -10,
   },
   headerTitle: {
     fontSize: 17,
@@ -223,14 +214,49 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     alignItems: 'center',
     marginHorizontal: 20,
+    marginTop: 5,
+  },
+  cardOwnerName: {
+    fontSize: 15,
+    fontWeight: '500',
+    color: colors.white,
+  },
+  cardType: {
+    fontSize: 13,
+    fontWeight: '500',
+    color: colors.white,
+  },
+  cardNumber: {
+    fontSize: 15,
+    fontWeight: '400',
+    color: colors.white,
+  },
+  cardExp: {
+    fontSize: 15,
+    fontWeight: '500',
+    color: colors.white,
+  },
+  cardNumberContainer: {
+    flexDirection: 'row',
+    marginHorizontal: 20,
+    marginTop: 8,
+  },
+  cardExpContainer: {
+    flexDirection: 'row',
+    marginHorizontal: 20,
     marginTop: 15,
   },
+  cardTypeContainer: {
+    flexDirection: 'row',
+    marginHorizontal: 20,
+    marginTop: 20,
+  },
   formContainer: {
-    marginTop: 30,
+    marginTop: 10,
     flexGrow: 1,
   },
   input: {
-    padding: 10,
+    padding: 7,
     backgroundColor: colors.grey5,
     borderBottomWidth: 0,
     borderEndColor: 'transparent',
@@ -245,7 +271,7 @@ const styles = StyleSheet.create({
   label: {
     fontSize: 17,
     fontWeight: '500',
-    marginBottom: 10,
+    marginBottom: 8,
     color: colors.black,
   },
   row: {
@@ -261,7 +287,7 @@ const styles = StyleSheet.create({
   toggle: {
     flexDirection: 'row',
     justifyContent: 'space-between',
-    alignItems: 'center',
+    alignItems: 'flex-start',
     paddingHorizontal: 20,
   },
   toggleText: {
@@ -272,20 +298,39 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
     justifyContent: 'center',
     backgroundColor: colors.primary,
-    paddingVertical: 20,
+    paddingVertical: 15,
   },
   buttonText: {
     color: 'white',
     fontSize: 17,
     fontWeight: '500',
   },
-  paymentLogo: {
-    backgroundColor: colors.grey5,
+  cardImage: {
+    width: 300,
+    height: 185,
+    marginHorizontal: 10,
+    marginTop: 20,
     borderRadius: 10,
+  },
+  cardRow: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    margin: 20,
+  },
+  addNewCardButton: {
+    backgroundColor: colors.primary1,
+    borderColor: colors.primary,
+    borderWidth: 1,
+    marginHorizontal: 20,
+    marginTop: 10,
     padding: 10,
-    width: 100,
-    height: 50,
-    justifyContent: 'center',
+    borderRadius: 10,
     alignItems: 'center',
+    flexDirection: 'row',
+    justifyContent: 'center',
+  },
+  addNewCardText: {
+    color: colors.primary,
+    marginLeft: 10,
   },
 });
